format code

+++ src/main/java/ar/com/hjg/pngj/IdatSet.java
@@ -102,7 +102,7 @@ public class IdatSet extends DeflatedChunksSet {
 
   private void unfilterRowNone(final int nbytes) {
     for (int i = 1; i <= nbytes; i++) {
-      rowUnfiltered[i] = (byte) (row[i]);
+        rowUnfiltered[i] = row[i];
     }
   }
 
@@ -120,7 +120,7 @@ public class IdatSet extends DeflatedChunksSet {
   private void unfilterRowSub(final int nbytes) {
     int i, j;
     for (i = 1; i <= imgInfo.bytesPixel; i++) {
-      rowUnfiltered[i] = (byte) (row[i]);
+        rowUnfiltered[i] = row[i];
     }
     for (j = 1, i = imgInfo.bytesPixel + 1; i <= nbytes; i++, j++) {
       rowUnfiltered[i] = (byte) (row[i] + rowUnfiltered[j]);
