+++ src/main/java/ar/com/hjg/pngj/PngHelperInternal.java
@@ -208,15 +208,15 @@ public final class PngHelperInternal {
 
   // / filters
   public static int filterRowNone(int r) {
-    return (int) (r & 0xFF);
+      return r & 0xFF;
   }
 
   public static int filterRowSub(int r, int left) {
-    return ((int) (r - left) & 0xFF);
+      return (r - left & 0xFF);
   }
 
   public static int filterRowUp(int r, int up) {
-    return ((int) (r - up) & 0xFF);
+      return (r - up & 0xFF);
   }
 
   public static int filterRowAverage(int r, int left, int up) {
