+++ src/main/java/ar/com/hjg/pngj/pixels/PixelsWriter.java
@@ -123,7 +123,7 @@ public abstract class PixelsWriter {
         break;
       case FILTER_SUB:
         for (i = 1; i <= bytesPixel; i++)
-          _rowf[i] = (byte) _rowb[i];
+          _rowf[i] = _rowb[i];
         for (j = 1, i = bytesPixel + 1; i <= bytesRow; i++, j++)
           _rowf[i] = (byte) (_rowb[i] - _rowb[j]);
         break;
